made affected zones controller (create and update methods not tested) and made some minor fixes

+++ Backend/LogicPersistence/LogicPersistence.Api/Models/DTOs/VolunteerDtos.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel.DataAnnotations;
+using System.Numerics;
 
 namespace LogicPersistence.Api.Models.DTOs;
 
@@ -86,7 +87,7 @@ public class VolunteerDisplayDto {
 	public string name { get; set; } = string.Empty;
 	public string surname { get; set; } = string.Empty;
 	public int prefix { get; set; }
-	public int phone_number { get; set; }
+	public BigInteger phone_number { get; set; }
 	public string address { get; set; } = string.Empty;
 	public string identification { get; set; } = string.Empty;
